chatbot-backend/requirements.txt
google-auth-oauthlib
google-auth
flask
mongoengine
pyjwt
pymongo[srv]
python-dotenv
flask_cors
flask-migrate==4.0.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dateutil